requirements.txt
# Core dependencies
gymnasium>=0.29.0
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0
omegaconf>=2.3.0
matplotlib>=3.7.0
tensorboard>=2.13.0
plotly>=5.15.0

# RL libraries
stable-baselines3>=2.0.0
sb3-contrib>=2.0.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
ruff>=0.0.280
pre-commit>=3.3.0

# Demo and visualization
streamlit>=1.25.0
gradio>=3.40.0

# Optional: logging and tracking
wandb>=0.15.0
mlflow>=2.5.0

# Optional: serving
fastapi>=0.100.0
uvicorn>=0.23.0
